# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi User Management Apis Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "usamauddin217@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of Users",
                "description": "Returns list of Users",
                "operationId": "get users",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Store Single new User",
                "description": "Returns Inserted User Resource",
                "operationId": "store user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get User information",
                "description": "Returns User Resource",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update existing User",
                "description": "Returns updated User Resource",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/users/permission": {
            "post": {
                "tags": [
                    "UserPermissions"
                ],
                "summary": "Store new User Permission",
                "description": "Returns User Permission Resource",
                "operationId": "storeUserPermisson",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "User ID": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "User Name": {
                                            "type": "string",
                                            "example": "Usamauddin"
                                        },
                                        "Permission ID": {
                                            "type": "string",
                                            "example": "2"
                                        },
                                        "Permission Name": {
                                            "type": "string",
                                            "example": "Update Employee Information"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserRequest": {
                "title": "User Form Request",
                "description": "user form request body data Validate",
                "required": [
                    "name",
                    "username",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "title": "User Fullname",
                        "description": "Fullname of the new User",
                        "type": "string",
                        "example": "Usamauddin"
                    },
                    "username": {
                        "title": "Username",
                        "description": "Username of the new User",
                        "type": "string",
                        "example": "usama44"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email of the new User",
                        "type": "string",
                        "example": "usama@gmail.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password of the new User",
                        "type": "string",
                        "example": "password@4422"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserRequest"
                }
            },
...
            "UserResource": {
                "title": "User Resource",
                "description": "User resource",
                "properties": {
                    "": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserResource"
                }
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "user_id": {
                        "title": "User ID",
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "User Fullname",
                        "description": "Fullname of the new User",
                        "type": "string",
                        "example": "Usamauddin"
                    },
                    "username": {
                        "title": "Username",
                        "description": "Username of the new User",
                        "type": "string",
                        "example": "usama44"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email of the new User",
                        "type": "string",
                        "example": "usama@gmail.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password of the new User",
                        "type": "string",
                        "example": "password@4422"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        }
    },
...
}